Abstracting out environment
Something that all of the code is going to want is some abstraction
around "what environment am I in?"

Right now, there's a lot of code that asks "is this LOCAL? PRODUCTION?"
Instead, this can be pushed down. And, it can be parameterized. For
example, there can be a variable called

ENVVAR

which lets us set which variable holds the name of the environment.
Then, we can have accessors for

IsLocalEnv

and

IsRemoteEnv

or similar, which lets us easily ask "where am I?"

For now, this only converts one function, but it demonstrates that it
can be cleaned up. Its a first step.